feat(opts): Add config option
