+++ ssltest/__main__.py
@@ -1,8 +1,9 @@
+from .Args import Args
 from .core.Script import run_script
 
 
 def main():
-    run_script()
+    run_script(Args.args, Args.parser)
 
 
 if __name__ == '__main__':
